+++ src/guards/RoleBasedGuard.tsx
@@ -1,4 +1,6 @@
 import { ReactNode } from "react";
+import { Navigate } from "react-router";
+import { toast, ToastContainer } from "react-toastify";
 
 type RoleBasedGuardProps = {
   children: ReactNode;
@@ -7,7 +9,7 @@ type RoleBasedGuardProps = {
 
 const useCurrentRole = () => {
   // current user role
-  const role = "admin";
+  const role = "superuser";
   return role;
 };
 
@@ -16,10 +18,17 @@ export default function RoleBasedGuard({
   children,
 }: RoleBasedGuardProps) {
   const currentRole = useCurrentRole();
+  const notify = () =>
+    toast.error(
+      "Permission Denied. You do not have permission to access this page"
+    );
 
   if (!accessibleRoles.includes(currentRole)) {
-    return alert(
-      "Permission Denied. You do not have permission to access this page"
+    notify();
+    return (
+      <>
+        <Navigate to={"/module/settings/student-type"} />;
+      </>
     );
   }